fix(Solution): Updated Neuroglia packages for solution, thus resolving a bug with command line escaping in Unix when evaluating inline escaped json fix(Worker): Fixed the InjectStateProcessor to evaluate the defined data before injecting it into the state's fix(Worker): Fixed the OpenApiFunctionProcessor to use the EvaluateObjectAsync IWorkflowRuntimeContext extension when building object, instead of using its own routine
