+++ src/apis/management/Synapse.Apis.Management.Http/Synapse.Apis.Management.Http.csproj
@@ -28,7 +28,7 @@
     <PackageReference Include="Microsoft.AspNetCore.Mvc.NewtonsoftJson" Version="6.0.10" />
     <PackageReference Include="Microsoft.AspNetCore.OData" Version="8.0.11" />
     <PackageReference Include="Microsoft.AspNetCore.OData.NewtonsoftJson" Version="8.0.4" />
-    <PackageReference Include="Neuroglia.Mediation.AspNetCore" Version="2.0.3" />
+    <PackageReference Include="Neuroglia.Mediation.AspNetCore" Version="2.0.4" />
     <PackageReference Include="Swashbuckle.AspNetCore.SwaggerGen" Version="6.4.0" />
     <PackageReference Include="Swashbuckle.AspNetCore.SwaggerUI" Version="6.4.0" />
   </ItemGroup>
