+++ src/apps/Synapse.Worker/Synapse.Worker.csproj
@@ -40,9 +40,9 @@
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
     </PackageReference>
     <PackageReference Include="Neuroglia.AsyncApi.Client.All" Version="2.1.0.6" />
-    <PackageReference Include="Neuroglia.Data.Expressions.JQ" Version="2.0.3" />
-    <PackageReference Include="Neuroglia.Eventing.CloudEvents" Version="2.0.3" />
-    <PackageReference Include="Neuroglia.Serialization.NewtonsoftJson" Version="2.0.3" />
+    <PackageReference Include="Neuroglia.Data.Expressions.JQ" Version="2.0.4" />
+    <PackageReference Include="Neuroglia.Eventing.CloudEvents" Version="2.0.4" />
+    <PackageReference Include="Neuroglia.Serialization.NewtonsoftJson" Version="2.0.4" />
     <PackageReference Include="protobuf-net.Grpc" Version="1.0.177" />
     <PackageReference Include="protobuf-net.Grpc.ClientFactory" Version="1.0.177" />
     <PackageReference Include="Simple.OData.Client" Version="5.26.0" />
