+++ src/core/Synapse.Application/Synapse.Application.csproj
@@ -27,10 +27,10 @@
     <PackageReference Include="CloudNative.CloudEvents.NewtonsoftJson" Version="2.5.0" />
     <PackageReference Include="Microsoft.AspNetCore.OData" Version="8.0.11" />
     <PackageReference Include="Microsoft.OData.ModelBuilder" Version="1.0.9" />
-    <PackageReference Include="Neuroglia.AspNetCore.JsonPatch" Version="2.0.3" />
-    <PackageReference Include="Neuroglia.Data.DistributedCache" Version="2.0.3" />
-    <PackageReference Include="Neuroglia.Data.EventSourcing" Version="2.0.3" />
-    <PackageReference Include="Neuroglia.Mediation.FluentValidation" Version="2.0.3" />
+    <PackageReference Include="Neuroglia.AspNetCore.JsonPatch" Version="2.0.4" />
+    <PackageReference Include="Neuroglia.Data.DistributedCache" Version="2.0.4" />
+    <PackageReference Include="Neuroglia.Data.EventSourcing" Version="2.0.4" />
+    <PackageReference Include="Neuroglia.Mediation.FluentValidation" Version="2.0.4" />
     <PackageReference Include="System.Linq.Async" Version="6.0.1" />
   </ItemGroup>
 </Project>
