+++ src/core/Synapse.Domain/Synapse.Domain.csproj
@@ -20,9 +20,9 @@
     <DebugType>embedded</DebugType>
   </PropertyGroup>
   <ItemGroup>
-    <PackageReference Include="Neuroglia.AspNetCore.JsonPatch" Version="2.0.3" />
-    <PackageReference Include="Neuroglia.Data" Version="2.0.3" />
-    <PackageReference Include="Neuroglia.Data.Kubernetes" Version="2.0.3" />
+    <PackageReference Include="Neuroglia.AspNetCore.JsonPatch" Version="2.0.4" />
+    <PackageReference Include="Neuroglia.Data" Version="2.0.4" />
+    <PackageReference Include="Neuroglia.Data.Kubernetes" Version="2.0.4" />
     <PackageReference Include="Semver" Version="2.2.0" />
   </ItemGroup>
   <ItemGroup>
