+++ src/core/Synapse.Integration/Synapse.Integration.csproj
@@ -32,10 +32,10 @@
   <ItemGroup>
     <PackageReference Include="Microsoft.AspNetCore.Http.Features" Version="5.0.17" />
     <PackageReference Include="Microsoft.AspNetCore.JsonPatch" Version="6.0.10" />
-    <PackageReference Include="Neuroglia.Serialization.Dynamic" Version="2.0.3" />
-    <PackageReference Include="Neuroglia.Serialization.Json" Version="2.0.3" />
-    <PackageReference Include="Neuroglia.Serialization.NewtonsoftJson" Version="2.0.3" />
-    <PackageReference Include="Neuroglia.Serialization.Protobuf" Version="2.0.3" />
+    <PackageReference Include="Neuroglia.Serialization.Dynamic" Version="2.0.4" />
+    <PackageReference Include="Neuroglia.Serialization.Json" Version="2.0.4" />
+    <PackageReference Include="Neuroglia.Serialization.NewtonsoftJson" Version="2.0.4" />
+    <PackageReference Include="Neuroglia.Serialization.Protobuf" Version="2.0.4" />
     <PackageReference Include="ServerlessWorkflow.Sdk" Version="0.8.3" />
     <PackageReference Include="System.ServiceModel.Primitives" Version="4.10.0" />
   </ItemGroup>
