+++ src/dashboard/Synapse.Dashboard/Synapse.Dashboard.csproj
@@ -22,13 +22,13 @@
     <PackageReference Include="Microsoft.AspNetCore.SignalR.Client" Version="6.0.10" />
     <PackageReference Include="Microsoft.AspNetCore.SignalR.Protocols.NewtonsoftJson" Version="6.0.10" />
     <PackageReference Include="Microsoft.OpenApi.Readers" Version="1.4.1" />
-    <PackageReference Include="Neuroglia.Blazor.Dagre" Version="2.0.3" />
-    <PackageReference Include="Neuroglia.Core" Version="2.0.3" />
-    <PackageReference Include="Neuroglia.Data.Flux" Version="2.0.3" />
-    <PackageReference Include="Neuroglia.Data.Flux.ReduxDevTools" Version="2.0.3" />
-    <PackageReference Include="Neuroglia.Data.OData" Version="2.0.3" />
-    <PackageReference Include="Neuroglia.Data.Schemas" Version="2.0.3" />
-    <PackageReference Include="Neuroglia.Mapping" Version="2.0.3" />
+    <PackageReference Include="Neuroglia.Blazor.Dagre" Version="2.0.4" />
+    <PackageReference Include="Neuroglia.Core" Version="2.0.4" />
+    <PackageReference Include="Neuroglia.Data.Flux" Version="2.0.4" />
+    <PackageReference Include="Neuroglia.Data.Flux.ReduxDevTools" Version="2.0.4" />
+    <PackageReference Include="Neuroglia.Data.OData" Version="2.0.4" />
+    <PackageReference Include="Neuroglia.Data.Schemas" Version="2.0.4" />
+    <PackageReference Include="Neuroglia.Mapping" Version="2.0.4" />
     <PackageReference Include="OData.QueryBuilder" Version="2.9.7" />
     <PackageReference Include="Semver" Version="2.2.0" />
   </ItemGroup>
